+++ crates/selector4nix-core/src/application/usecase/mod.rs
@@ -2,3 +2,4 @@ pub mod dashboard;
 pub mod derivation;
 pub mod nar_file;
 pub mod nar_info;
+pub mod substituter;