+++ crates/selector4nix-web/src/dashboard/mod.rs
@@ -2,12 +2,14 @@ mod cache;
 mod configuration;
 mod overview;
 mod statics;
+mod substituters;
 mod transferring;
 
 pub use cache::get_cache_page;
 pub use configuration::get_configuration_page;
 pub use overview::get_overview_page;
 pub use statics::get_static_asset;
+pub use substituters::{post_add_substituter, post_disable_substituter, post_enable_substituter};
 pub use transferring::get_transferring_page;
 
 use std::sync::LazyLock;